refactor(ducklake): rename data_modeling to duck copy workflow to be more specific
