+++ posthog/temporal/data_modeling/run_workflow.py
@@ -45,6 +45,7 @@ from posthog.temporal.common.heartbeat import Heartbeater
 from posthog.temporal.common.logger import get_logger
 from posthog.temporal.data_imports.util import prepare_s3_files_for_querying
 from posthog.temporal.data_modeling.metrics import get_data_modeling_finished_metric
+from posthog.temporal.ducklake.ducklake_copy_data_modeling_workflow import DuckLakeCopyDataModelingWorkflow
 from posthog.temporal.utils import DataModelingDuckLakeCopyInputs, DuckLakeCopyModelInput
 
 from products.data_warehouse.backend.data_load.create_table import create_table_from_saved_query
@@ -1600,9 +1601,10 @@ class RunWorkflow(PostHogWorkflow):
                 job_id=job_id,
                 models=len(self.ducklake_copy_inputs.models),
             )
+            # Start DuckLake copy workflow as a child (fire-and-forget)
             await temporalio.workflow.start_child_workflow(
-                workflow="ducklake-copy.data-modeling",
-                arg=dataclasses.asdict(self.ducklake_copy_inputs),
+                DuckLakeCopyDataModelingWorkflow.run,
+                self.ducklake_copy_inputs,
                 id=f"ducklake-copy-data-modeling-{job_id}",
                 task_queue=settings.DUCKLAKE_TASK_QUEUE,
                 parent_close_policy=ParentClosePolicy.ABANDON,
